refactor: move health-check component in api package
