+++ gravitee-apim-console-webui/src/management/api/health-check/api-health-check.component.html
@@ -20,9 +20,9 @@
     Health check configuration let you enable or disable health check service, set a frequency rate, choose an health relative path and
     provide some assertions to control endpoint API's health.
   </p>
-  <button class="api-proxy-health-check__header__go-to-dashboard" mat-raised-button type="button" [routerLink]="'../healthcheck-dashboard'">
+  <a class="api-proxy-health-check__header__go-to-dashboard" mat-raised-button type="button" [routerLink]="'../healthcheck-dashboard'">
     Go to health check dashboard
-  </button>
+  </a>
 </div>
 
 <gio-banner-info class="api-proxy-health-check__banner-info">
