+++ gravitee-apim-console-webui/src/management/api/health-check/api-health-check.component.scss
